Refactored Package common to com
Imported sun.com.tool.utils.Pair in the project under java.com.utils.Pair as it is not a JRE class

Converted the Majority of Tests for components from simple "eye-to-eye" test to ScalaTest and AkkaTest

NOTE: As for a problem with the configuration of IntelliJ and/or gradle, or whatever the reason may be, all the tests are in the arianna_main module and not in the arianna_test module. The problem is that the IDE is unable to find the test classes... Hopefully we will be able to move those tests where they properly belong.
Also, being a cross-language project (scala + java, and Kotlin for the Android Application, located in a different Repository -- see Project Description), we were unable to move from gradle to SBT as some Java-based library where nowhere to be found on SBT repositories.